Add blanket impl and trait bound
